TaskSchedule/requirements.txt
fastapi==0.111.1
# uvicorn[standard]==0.30.4
pyyaml==6.0.1
geopy==2.4.1